+++ applications/fishing-map/src/features/workspace/common/InfoModal.tsx
@@ -8,8 +8,9 @@ import Tabs, { Tab } from '@globalfishingwatch/ui-components/dist/tabs'
 import Modal from '@globalfishingwatch/ui-components/dist/modal'
 import { IconButton } from '@globalfishingwatch/ui-components'
 import { DatasetStatus } from '@globalfishingwatch/api-types/dist'
+import { GeneratorType } from '@globalfishingwatch/layer-composer/dist/generators'
 import { getDatasetDescriptionTranslated } from 'features/i18n/utils'
-import { getDatasetLabel } from 'features/datasets/datasets.utils'
+import { getDatasetLabel, hasDatasetConfigVesselData } from 'features/datasets/datasets.utils'
 import { isGFWUser } from 'features/user/user.slice'
 import styles from './InfoModal.module.css'
 
@@ -29,7 +30,14 @@ const InfoModal = ({ dataview, onClick, className, onModalStateChange }: InfoMod
   const tabs = useMemo(() => {
     const uniqDatasets = dataview.datasets ? uniqBy(dataview.datasets, (dataset) => dataset.id) : []
     return uniqDatasets.flatMap((dataset) => {
-      if (dataview.config?.datasets && !dataview.config?.datasets?.includes(dataset.id)) {
+      if (dataview.config?.type === GeneratorType.Track) {
+        const datasetConfig = dataview.datasetsConfig?.find(
+          (datasetConfig) => datasetConfig.datasetId === dataset.id
+        )
+        if (!datasetConfig) return []
+        const hasDatasetVesselId = hasDatasetConfigVesselData(datasetConfig)
+        if (!hasDatasetVesselId) return []
+      } else if (dataview.config?.datasets && !dataview.config?.datasets?.includes(dataset.id)) {
         return []
       }
       const description = getDatasetDescriptionTranslated(dataset)
@@ -115,7 +123,7 @@ const InfoModal = ({ dataview, onClick, className, onModalStateChange }: InfoMod
       />
     ) : null
   }
-
+  const hasLongTitleTab = tabs.some((tab) => tab.title.length > 30)
   return (
     <Fragment>
       <IconButton
@@ -128,7 +136,7 @@ const InfoModal = ({ dataview, onClick, className, onModalStateChange }: InfoMod
         tooltipPlacement="top"
         onClick={handleClick}
       />
-      {tabs && tabs.length > 0 && (
+      {tabs && tabs.length > 0 && modalInfoOpen && (
         <Modal
           title={isSingleTab ? tabs[0].title : dataview.name}
           isOpen={modalInfoOpen}
@@ -142,7 +150,7 @@ const InfoModal = ({ dataview, onClick, className, onModalStateChange }: InfoMod
               tabs={tabs}
               activeTab={activeTab?.id}
               onTabClick={(tab: Tab) => setActiveTab(tab)}
-              buttonSize="verybig"
+              buttonSize={hasLongTitleTab ? 'verybig' : 'default'}
             />
           )}
         </Modal>