+++ applications/fishing-map/src/features/workspace/vessels/VesselLayerPanel.tsx
@@ -250,21 +250,31 @@ function LayerPanel({ dataview }: LayerPanelProps): React.ReactElement {
         ) : (
           TitleComponent
         )}
-        <div className={cx('print-hidden', styles.actions, { [styles.active]: layerActive })}>
+        <div
+          className={cx('print-hidden', styles.actions, styles.hideUntilHovered, {
+            [styles.active]: layerActive,
+          })}
+        >
           <Fragment>
             {gfwUser && (
               <IconButton
                 icon="download"
                 tooltip={t('download.trackAction', 'Download vessel track')}
+                tooltipPlacement="top"
                 onClick={onDownloadClick}
                 size="small"
               />
             )}
             {layerActive && !infoLoading && TrackIconComponent}
-            {infoResource && InfoIconComponent}
             <Remove dataview={dataview} />
           </Fragment>
+          {infoResource && InfoIconComponent}
         </div>
+        <IconButton
+          icon="more"
+          className={cx('print-hidden', styles.shownUntilHovered)}
+          size="small"
+        />
       </div>
     </div>
   )