+++ applications/vessel-history/src/features/profile/components/activity/ActivityItem.tsx
@@ -3,9 +3,9 @@ import { RenderedEvent } from 'features/vessels/activity/vessels-activity.select
 import { EventTypeVoyage, RenderedVoyage, Voyage } from 'types/voyage'
 import ActivityEvent from './ActivityEvent'
 import ActivityVoyage from './ActivityVoyage'
-
 interface EventProps {
   event: RenderedEvent | RenderedVoyage
+  highlighted?: boolean
   onInfoClick?: (event: RenderedEvent) => void
   onMapClick?: (event: RenderedEvent | Voyage) => void
   onToggleClick?: (event: RenderedVoyage) => void
@@ -13,6 +13,7 @@ interface EventProps {
 
 const ActivityItem: React.FC<EventProps> = ({
   event,
+  highlighted = false,
   onInfoClick = () => {},
   onMapClick = () => {},
   onToggleClick = () => {},
@@ -29,9 +30,10 @@ const ActivityItem: React.FC<EventProps> = ({
       {event.type !== EventTypeVoyage.Voyage && (
         <ActivityEvent
           event={event}
+          highlighted={highlighted}
           onMapClick={onMapClick}
           onInfoClick={onInfoClick}
-        ></ActivityEvent>
+          ></ActivityEvent>
       )}
     </Fragment>
   )