+++ packages/ui-components/src/tabs/index.ts
@@ -2,6 +2,6 @@ export { default } from './Tabs'
 
 export interface Tab {
   id: string
-  title: string
+  title: string | React.ReactElement
   content: React.ReactNode
 }