+++ applications/tile-inspector/CHANGELOG.md
@@ -1,5 +1,26 @@
 # tile-inspector
 
+## 0.1.99
+
+### Patch Changes
+
+- Updated dependencies [b88273847]
+  - @globalfishingwatch/layer-composer@10.0.2
+
+## 0.1.98
+
+### Patch Changes
+
+- Updated dependencies [24642e15]
+  - @globalfishingwatch/ui-components@12.4.0
+
+## 0.1.97
+
+### Patch Changes
+
+- Updated dependencies [39c43baa]
+  - @globalfishingwatch/ui-components@12.3.1
+
 ## 0.1.96
 
 ### Patch Changes