+++ packages/i18n-labels/CHANGELOG.md
@@ -1,6 +1,13 @@
 # @globalfishingwatch/i18n-labels
 
+## 1.1.2
+
+### Patch Changes
+
+- add other vessels dataset translations
+
 ## 1.1.1
+
 ### Patch Changes
 
 - update datasets