+++ packages/ui-components/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @globalfishingwatch/ui-components
 
+## 12.4.0
+
+### Minor Changes
+
+- 24642e15: ACTIVITY tab - add a red dot icon in the events that are matching highlight criteria
+
+## 12.3.1
+
+### Patch Changes
+
+- 39c43baa: Minor tweaks in Switch styles
+
 ## 12.3.0
 
 ### Minor Changes